+++ src/Language/Haskell/TH/Instances.hs
@@ -387,6 +387,9 @@ instance Quasi m => Quasi (ReaderT r m) where
 #if MIN_VERSION_template_haskell(2,12,0)
   qAddForeignFile a b = MTL.lift $ qAddForeignFile a b
 #endif
+#if MIN_VERSION_template_haskell(2,13,0)
+  qAddCorePlugin      = MTL.lift . qAddCorePlugin
+#endif
 
 instance (Quasi m, Monoid w) => Quasi (WriterT w m) where
   qNewName            = MTL.lift . qNewName
@@ -419,6 +422,9 @@ instance (Quasi m, Monoid w) => Quasi (WriterT w m) where
 #endif
 #if MIN_VERSION_template_haskell(2,12,0)
   qAddForeignFile a b = MTL.lift $ qAddForeignFile a b
+#endif
+#if MIN_VERSION_template_haskell(2,13,0)
+  qAddCorePlugin      = MTL.lift . qAddCorePlugin
 #endif
 
 instance Quasi m => Quasi (StateT s m) where
@@ -453,6 +459,9 @@ instance Quasi m => Quasi (StateT s m) where
 #if MIN_VERSION_template_haskell(2,12,0)
   qAddForeignFile a b = MTL.lift $ qAddForeignFile a b
 #endif
+#if MIN_VERSION_template_haskell(2,13,0)
+  qAddCorePlugin      = MTL.lift . qAddCorePlugin
+#endif
 
 instance (Quasi m, Monoid w) => Quasi (RWST r w s m) where
   qNewName            = MTL.lift . qNewName
@@ -485,6 +494,9 @@ instance (Quasi m, Monoid w) => Quasi (RWST r w s m) where
 #endif
 #if MIN_VERSION_template_haskell(2,12,0)
   qAddForeignFile a b = MTL.lift $ qAddForeignFile a b
+#endif
+#if MIN_VERSION_template_haskell(2,13,0)
+  qAddCorePlugin      = MTL.lift . qAddCorePlugin
 #endif
 
 #if MIN_VERSION_base(4,7,0) && defined(LANGUAGE_DeriveDataTypeable) && __GLASGOW_HASKELL__ < 710