+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### next [????.??.??]
+* Implement `qAddCorePlugin` for `Quasi` instances
+
 ### 0.13.4 [2017.07.26]
 * Implement `qAddForeignFile` for `Quasi` instances